cluster: disclose the forged-liveness half of the restart replay
The arming-site comment told an operator that an archived epoch-bearing
frame re-arms the downgrade latch against fresh post-restart state, and
that rotating the PSK before restarting is the complete recovery. It did
not say what else that frame does. The same frame is ADMITTED, so
admitFrame goes on to refresh lastSeen and call handlePeerHeartbeat: a
peer that is dead looks alive for as long as the replay continues, and
that liveness feeds election. One frame per dead-peer interval sustains
it. A reader would take the existing paragraph as the whole cost.

Measured rather than assumed: master has no admitAuthed at all, so the
bare replay ring admits the same frame after the same receiver restart,
and TestHeartbeatRestartStillAcceptsGenuinePeer_5086 requires exactly
that — a genuine peer reboot must be accepted. So this is pre-existing
rather than introduced here, the epoch floor strictly improves on master
once armed, and the honest fix is disclosure, not a behavioural change
that would trade a fail-open for a fail-closed. Both the code comment and
README's residual list now carry it.

Widening the comment pushed heartbeat.go past the 2000-line modularity
gate, which the previous commit on this branch had already taken from
1988 to 2050 without re-measuring. Two extractions restore compliance,
and both are the right cut independently of the count:
heartbeat_reject_warn.go takes the rate limiter added this round, and
heartbeat_epoch_admit.go takes admitAuthed, epochSessionAdmissible and
bindEpochSession with their rationale. heartbeat.go keeps the wire format
and the sender/receiver lifecycle; nothing in the extracted file touches
sockets. heartbeat.go is now 1595 and the refactoring audit reports WATCH
rather than REFACTOR.

go build ./... 0; go test -count=1 ./pkg/cluster/... 0; gofmt clean;
scripts/refactoring-audit.sh clean.
